fix(sidecar): skip TLS config for plain HTTP URLs, harden port validation
Skip building TLS configuration when the target URL scheme is http,
preventing crashes from missing CA cert files in local development.
For https URLs, insecure_skip_verify continues to bypass CA cert
requirements as before.

Tighten ResolvePort to require http/https scheme, a hostname, and an
explicit port in BaseURL (needed for Kubernetes probe alignment).
Change SidecarConfig.Port to int32 to eliminate gosec integer overflow
warnings at the k8s probe-port assignment site.

Harden start_sidecar.sh temp file creation with mktemp, restrictive
permissions, cleanup trap, and fail-fast error handling.

+++ internal/eval_hub/runtimes/k8s/job_builders.go
@@ -174,7 +174,7 @@ func buildJob(cfg *jobConfig) (*batchv1.Job, error) {
 	}
 	probePort := int32(config.DefaultSidecarPort)
 	if cfg.sidecarConfig != nil && cfg.sidecarConfig.Port != 0 {
-		probePort = int32(cfg.sidecarConfig.Port)
+		probePort = cfg.sidecarConfig.Port
 	}
 
 	// Sidecar is added as an init container with restartPolicy=Always to be
